+++ packages/react-lua-hooks/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+* fix `usePrevious` and `usePreviousDistinct` hooks ([#5](https://github.com/seaofvoices/react-lua-hooks/pull/5))
+
 ## 0.1.0
 
 * Initial version